Move the biome structure loading code into the same file as saving.
