+++ chunky/src/java/se/llbit/chunky/renderer/scene/biome/BiomeStructure.java
@@ -22,110 +22,8 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
 
   static void registerDefaults() {
     //TODO: create a plugin api interface for registering implementations, and move this to that
-    BiomeStructure.register("TRIVIAL_3D", new Factory() {
-      @Override
-      public BiomeStructure create() {
-        return new Trivial3dBiomeStructureImpl();
-      }
-
-      @Override
-      public BiomeStructure load(@NotNull DataInputStream in) throws IOException {
-        /*
-         * Stored as:
-         * (int) size
-         * (int) x, y, z
-         * (long) Length of present bitset in longs
-         * (BitSet as longs) Present values bitset
-         * (int) number of values stored
-         * (float[][]) The internal data of each packed x,y,z position
-         */
-
-        Trivial3dBiomeStructureImpl impl = new Trivial3dBiomeStructureImpl();
-        int size = in.readInt();
-        for (int i = 0; i < size; i++) {
-          int x = in.readInt();
-          int y = in.readInt();
-          int z = in.readInt();
-
-          long[] longs = new long[in.readInt()];
-          for (int bitsetIdx = 0; bitsetIdx < longs.length; bitsetIdx++) {
-            longs[bitsetIdx] = in.readLong();
-          }
-
-          BitSet presentValues = BitSet.valueOf(longs);
-
-          int count = in.readInt();
-          float[][] floats = new float[count][];
-          for (int idx = 0; idx < count; idx++) {
-            if(presentValues.get(idx)) {
-              float[] farray = new float[3];
-              farray[0] = in.readFloat();
-              farray[1] = in.readFloat();
-              farray[2] = in.readFloat();
-              floats[idx] = farray;
-            }
-          }
-          impl.setCube(x, y, z, floats);
-        }
-        return impl;
-      }
-
-      @Override
-      public boolean is3d() {
-        return true;
-      }
-    });
-
-    BiomeStructure.register("TRIVIAL_2D", new Factory() {
-      @Override
-      public BiomeStructure create() {
-        return new Trivial2dBiomeStructureImpl();
-      }
-
-      /**
-       * Stored as:
-       * (int) size
-       * (int) long packed key
-       * (long) Length of present bitset in longs
-       * (BitSet as longs) Present values bitset
-       * (int) number of values stored
-       * (float[][]) The internal data of each packed x,z position
-       */
-      @Override
-      public BiomeStructure load(@NotNull DataInputStream in) throws IOException {
-        Trivial2dBiomeStructureImpl impl = new Trivial2dBiomeStructureImpl();
-        int size = in.readInt();
-        for (int i = 0; i < size; i++) {
-          long key = in.readLong();
-
-          long[] longs = new long[in.readInt()];
-          for (int bitsetIdx = 0; bitsetIdx < longs.length; bitsetIdx++) {
-            longs[bitsetIdx] = in.readLong();
-          }
-
-          BitSet presentValues = BitSet.valueOf(longs);
-
-          int count = in.readInt();
-          float[][] floats = new float[count][];
-          for (int idx = 0; idx < count; idx++) {
-            if(presentValues.get(idx)) {
-              float[] farray = new float[3];
-              farray[0] = in.readFloat();
-              farray[1] = in.readFloat();
-              farray[2] = in.readFloat();
-              floats[idx] = farray;
-            }
-          }
-          impl.setCube(key, floats);
-        }
-        return impl;
-      }
-
-      @Override
-      public boolean is3d() {
-        return false;
-      }
-    });
+    BiomeStructure.register("TRIVIAL_3D", new Trivial3dBiomeStructure());
+    BiomeStructure.register("TRIVIAL_2D", new Trivial2dBiomeStructure());
   }
 
   /**
@@ -133,7 +31,7 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
    * BiomeStructure implementation
    *
    * @param impl The implementation to load the legacy implementation into
-   * @param in The serialised legacy data in an input stream
+   * @param in   The serialised legacy data in an input stream
    * @return The newly constructed {@link BiomeStructure} of the specified implementation
    */
   static BiomeStructure loadLegacy(Factory impl, DataInputStream in) throws IOException {
@@ -144,7 +42,7 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
       int chunkZ = in.readInt();
       for (int localX = 0; localX < Chunk.X_MAX; localX++) {
         for (int localZ = 0; localZ < Chunk.Z_MAX; localZ++) {
-          biomeStructure.set(chunkX * Chunk.X_MAX + localX, 0, chunkZ * Chunk.Z_MAX + localZ, new float[] {
+          biomeStructure.set(chunkX * Chunk.X_MAX + localX, 0, chunkZ * Chunk.Z_MAX + localZ, new float[]{
             in.readFloat(),
             in.readFloat(),
             in.readFloat()
@@ -158,6 +56,7 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
 
   /**
    * Register a new {@link Factory}
+   *
    * @param key The key to register the factory under <b>(MUST BE UNIQUE)</b>
    * @return Whether the supplied factory overwrote an existing one
    */
@@ -172,7 +71,7 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
   @NotNull
   static Factory get(@NotNull String key) throws NullPointerException {
     Factory factory = REGISTRY.get(key);
-    if(factory == null) {
+    if (factory == null) {
       Log.warnf("Implementation %s does not exist, using the default: %s", key, DEFAULT_IMPLEMENTATION);
       return REGISTRY.get(DEFAULT_IMPLEMENTATION);
     }
@@ -206,7 +105,7 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
      * (used only for biome blending, does not need to be saved)
      */
     default Position2IntStructure createIndexStructure() {
-      if(is3d()) {
+      if (is3d()) {
         return new Position3d2IntPackedArray();
       } else {
         return new Position2d2IntPackedArray();
